Unify tests into one file
Renamed to homepage to be consistent with other apps
Created common-scenarios which runs all 'logged in' tests

Remainder of tests need to be integrated into common-scenarios,
but currently they require the user to be logged out

Added verification that secondary (jekins1) credential is present
Trying to fix issue where jenkins company could be deleted
